Change extensionKind from ui+workspace to workspace (#4)
If you install from Marketplace, it will probably be installed as a ui, but in that state it did not work as intended. However, when installed from a vsix file using the code command on WSL, it worked fine.

This change may not fix all the issues, but it should fix some.